Append to gitignore and create envs on all commands
