+++ src/commands/init.ts
@@ -1,6 +1,7 @@
 import {Command, flags} from '@oclif/command'
 import {vars} from '../vars'
 import {AppendToGitignoreService} from '../services/append-to-gitignore-service'
+import {WriteEnvsService} from '../services/write-envs-service'
 
 export default class Init extends Command {
   static description = 'initialize .env.me and .env.project'
@@ -11,7 +12,12 @@ export default class Init extends Command {
   async run() {
     const {args, flags} = this.parse(Init)
 
-    const fs = require('fs')
+    // 1. create gitignore
+    new AppendToGitignoreService().run()
+
+    // 2. create envs
+    new WriteEnvsService().run()
+
     const path = require('path')
     const axios = require('axios')
     const dotenv = require('dotenv')
@@ -19,53 +25,7 @@ export default class Init extends Command {
     const signale = require('signale')
     const prompts = require('prompts')
 
-    const dir = process.cwd()
-
-    const dotenvFilename = '.env'
-    const dotenvData = 'KEY=value'
-
-    const meUid = crypto.randomBytes(32).toString('hex') // 64 length
-    const meFilename = '.env.me'
-    const meData = 'DOTENV_ME=me_' + meUid
-
-    const projectUid = crypto.randomBytes(32).toString('hex') // 64 length
-    const projectFilename = '.env.project'
-    const projectName = dir.split("/")[dir.split("/").length-1]
-    const projectData = 'DOTENV_PROJECT=prj_' + projectUid + "\n" + 'DOTENV_PROJECT_NAME=' + projectName
-
-    function logExisting(filename) {
-      signale.complete(filename + ' exists')
-    }
-
-    function logCreating(filename) {
-      signale.complete(filename + ' created')
-    }
-
-		// 1. create gitignore
-		new AppendToGitignoreService().run()
-
-    if (fs.existsSync(dotenvFilename)) {
-      logExisting(dotenvFilename)
-    } else {
-      fs.writeFileSync(dotenvFilename, dotenvData)
-      logCreating(dotenvFilename)
-    }
-
-    if (fs.existsSync(meFilename)) {
-      logExisting(meFilename)
-    } else {
-      fs.writeFileSync(meFilename, meData)
-      logCreating(meFilename)
-    }
-
-    if (fs.existsSync(projectFilename)) {
-      logExisting(projectFilename)
-    } else {
-      fs.writeFileSync(projectFilename, projectData)
-      logCreating(projectFilename)
-    }
-
-
+       
     const promptForShortCode = async () => {
       const response = await prompts({
         type: 'text',
